use wider types for pointers, so that it's easier to visually parse the binary output ;)

+++ test/generate.js
@@ -10,7 +10,7 @@ var schema = {
   simpler: [
     {name: 'foo', position: 0, direct: {type: 'u8'}},
     {name: 'bar', position: 1, direct: {type: 'u32'}},
-    {name: 'name', position: 5, direct: {type: 'u8'}, pointed: {type: 'string_u8'}}
+    {name: 'name', position: 5, direct: {type: 'u32'}, pointed: {type: 'string_u8'}}
   ]
 }
